Norminette.

+++ src/move.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   move.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: josantos <josantos@student.42.fr>          +#+  +:+       +#+        */
+/*   By: msousa <mlrcbsousa@gmail.com>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/10 00:13:00 by msousa            #+#    #+#             */
-/*   Updated: 2022/05/11 00:33:32 by josantos         ###   ########.fr       */
+/*   Updated: 2022/05/12 02:10:21 by msousa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ static void	move_down(t_app *self)
 
 	player = self->player;
 	delta = point(cos(player->angle), sin(player->angle));
-	delta = point_multiply(-MOVE, delta); // or point_subtract with +MOVE
+	delta = point_multiply(-MOVE, delta);
 	delta = point_add(player->p, delta);
 	if (is_element_bounded_and_not_wall(self->settings, delta))
 		player->p = delta;
@@ -52,7 +52,7 @@ static void	move_left(t_app *self)
 {
 	t_point		delta;
 	t_player	*player;
-	int			angle;
+	double		angle;
 
 	player = self->player;
 	angle = trim(player->angle - PI / 2);
@@ -67,7 +67,7 @@ static void	move_right(t_app *self)
 {
 	t_point		delta;
 	t_player	*player;
-	int			angle;
+	double		angle;
 
 	player = self->player;
 	angle = trim(player->angle + PI / 2);
